Pull resource limits into a separate source file.
