+++ tests/test_cfg_cache_ctrl.py
@@ -8,7 +8,7 @@ from unittest.mock import MagicMock
 import pytest
 
 from datacube_ows.ogc_utils import ConfigException
-from datacube_ows.ows_configuration import CacheControlRules
+from datacube_ows.resource_limits import CacheControlRules
 
 
 @pytest.fixture
